#64 -- Render speed
Removed functions for bool display. Not ternary. Trying to improve render.

Render speed is due to number of components showing in one action. Each
component is ~5ms, but have hundreds of components on one button click.

To improve speed, reduce number of components in edit mode. Two buttons
on side, with workflow of clicking element to edit or delete in focus